pyproject.toml
[project]
name = "book-review-fastapi"
version = "0.1.0"
description = "fastapi book review webservice"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.14.0",
    "asyncpg>=0.30.0",
    "fastapi[all]>=0.115.6",
    "passlib>=1.7.4",
    "pydantic-settings>=2.7.1",
    "ruff>=0.9.1",
    "sqlmodel>=0.0.22",
]
